refactor(plugin/aruba): move generic impls from port to adapter pkg
