+++ foundation/plugin/aruba/pkg/adapter/generic/delegated/delegated_test.go
