+++ foundation/plugin/aruba/pkg/adapter/generic/delegated/zz_mock_identifiable_test.go
