+++ foundation/plugin/aruba/pkg/port/converter/converter.go
@@ -26,25 +26,3 @@ type Converter[S any, A any] interface {
 	// FromArubaToSECA converts an Aruba resource to a SECA resource.
 	FromArubaToSECA(from A) (S, error)
 }
-
-// GenericConverter is a generic implementation of the Converter interface.
-//
-// It uses two functions to perform the conversions.
-type GenericConverter[S any, A any] struct {
-	fromSECAToAruba ConvertSECAToArubaFunc[S, A]
-	fromArubaToSECA ConvertArubaToSECAFunc[S, A]
-}
-
-// FromSECAToAruba converts a SECA resource to an Aruba resource using the
-// provided conversion function.
-func (c *GenericConverter[S, A]) FromSECAToAruba(from S) (A, error) {
-	return c.fromSECAToAruba(from)
-}
-
-// FromArubaToSECA converts an Aruba resource to a SECA resource using the
-// provided conversion function.
-func (c *GenericConverter[S, A]) FromArubaToSECA(from A) (S, error) {
-	return c.fromArubaToSECA(from)
-}
-
-var _ Converter[float64, int] = (*GenericConverter[float64, int])(nil)
